draughts, follower, smooth2, object tracker update
git-svn-id: https://segomo.elka.pw.edu.pl/svn/mrrocpp/base/trunk@2945 4c39dba3-d726-0410-8099-c98d187771e6

+++ src/ecp/irp6_on_track/ecp_vis_ib_eih_follower_irp6ot.cc
@@ -144,7 +144,7 @@ bool ecp_vis_ib_eih_follower_irp6ot::next_step_without_constraints() {
 	lib::VSP_REPORT vsp_report = vsp_fradia->from_vsp.vsp_report;
 	if (vsp_report == lib::VSP_REPLY_OK) {
 
-		if (fabs(u[0]) < 20 && fabs(u[1]) < 20 && v[0] <= v_stop[0] & v[1] <= v_stop[1]) {
+		if (fabs(u[0]) < 20 && fabs(u[1]) < 20 && v[0] <= v_stop[0] && v[1] <= v_stop[1]) {
 			if (z_stop) {
 				printf("#################################### koniec sledzenia ##################################\n");
 				flushall();
